+++ conf/production/db.conf.php
@@ -1,26 +1,12 @@
 <?php
-/** \file
- *  \brief      Database configurations.
- *  \copyright	Copyright (c) 2007-2016 Fernando Val
- *  \warning    Este arquivo é parte integrante do framework e não pode ser omitido.
- */
 
-/**
- *  \defgroup dbcfg_production Configurações de acesso a banco de dados para o ambiente 'production'
- *  \ingroup dbcfg.
+/*
+ * Springy Framework Configuration File.
  *
- *  As entradas colocadas nesse arquivo serão aplicadas apenas ao ambiente 'production'.
+ * Used for "production" environment.
  *
- *  Seu sistema pode não possuir esse ambiente, então use-o como modelo para criação do arquivo de
- *  parâmetros de configuração para os ambientes que seu sistema possua.
- *
- *  Veja \link dbcfg Configurações de acesso a banco de dados \endlink para entender as entradas de configuração possíveis.
- *
- *  \see dbcfg
+ * If removed, only db.default.conf.php will be used.
  */
-/**@{*/
-
-/// Configurações para o ambiente de Produção
 $conf = [
     'round_robin' => [
         'type'        => 'memcached',
@@ -42,5 +28,3 @@ $conf = [
         'persistent'    => false,
     ],
 ];
-
-/**@}*/