+++ conf/production/mail.conf.php
@@ -1,26 +1,12 @@
 <?php
-/** \file
- *  \brief      Configurations for Springy\Mail class.
- *  \copyright  Copyright (c) 2007-2016 Fernando Val
- *  \warning    Este arquivo é parte integrante do framework e não pode ser omitido.
- */
 
-/**
- *  \defgroup emailcfg_production Configurações da classe de envio de email para o ambiente \c 'production'
- *  \ingroup emailcfg.
+/*
+ * Springy Framework Configuration File.
  *
- *  As entradas colocadas nesse arquivo serão aplicadas apenas ao ambiente 'production'.
+ * Used for "production" environment.
  *
- *  Seu sistema pode não possuir esse ambiente, então use-o como modelo para criação do arquivo de
- *  parâmetros de configuração para os ambientes que seu sistema possua.
- *
- *  Veja \link emailcfg Configurações da classe de envio de email \endlink para entender as entradas de configuração possíveis.
- *
- *  \see emailcfg
+ * If removed, only mail.default.conf.php will be used.
  */
-/**@{*/
-
-/// Configurações para o ambiente de Produção
 $conf = [
     'default_driver' => 'sendgrid-api',
     'mailers'        => [
@@ -29,5 +15,3 @@ $conf = [
         ],
     ],
 ];
-
-/**@}*/