+++ conf/production/template.conf.php
@@ -1,30 +1,13 @@
 <?php
-/** \file
- *  \brief      Configurations for Springy\Template class.
- *  \copyright  Copyright (c) 2007-2016 FVAL Consultoria e Informática Ltda.\n
- *  \warning    Este arquivo é parte integrante do framework e não pode ser omitido.
- */
 
-/**
- *  \addtogroup templatecfg Configurações da classe de templates
- *  \defgroup templatecfg_production Configurações da classe de templates para o ambiente 'production'
- *  \ingroup templatecfg.
+/*
+ * Springy Framework Configuration File.
  *
- *  As entradas colocadas nesse arquivo serão aplicadas apenas ao ambiente 'production'.
+ * Used for "production" environment.
  *
- *  Seu sistema pode não possuir esse ambiente, então use-o como modelo para criação do arquivo de
- *  parâmetros de configuração para os ambientes que seu sistema possua.
- *
- *  Veja \link templatecfg Configurações da classe de templates \endlink para entender as entradas de configuração possíveis.
- *
- *  \see templatecfg
+ * If removed, only template.default.conf.php will be used.
  */
-/**@{*/
-
-/// Configurações para o ambiente de Produção
 $conf = [
     'debug'       => false,
     'auto_reload' => false,
 ];
-
-/**@}*/